+++ Civi/ActionProvider/Action/Relationship/GetRelationship.php
@@ -28,7 +28,6 @@ class GetRelationship extends AbstractAction {
       $this->relationshipTypes[$relType['name_a_b']] = $relType['label_a_b'];
       $this->relationshipTypeIds[$relType['name_a_b']] = $relType['id'];
     }
-
   }
 
   /**